+++ common/src/main/rust/marten/src/lib.rs
@@ -1,4 +1,4 @@
 pub mod level;
 pub mod octree;
 
-pub type Real = f32;
+pub type Real = f64;
